Make colorscheme more consistent
Why:

* CursorColumn is now the same as CursorLine.
StatusLineTerm's are now linked with StatusLine's.
Pmenu and WildMenu now use the same highlighting and are now white text
on black background for better contrast & less frivolous colors.
Changed the `hi link` to `hi! link` settings. I'm not quite sure what it
does, but everyone else seems to be doing it, so maybe it's a good
thing? I suspect it is similar to `au` v `au!` so probabbly some
recursion stuff???

This change addresses the need by:

* Me