Create options file for markdown & move settings
Why:

* The calling of Tabularize for '|' inserts made it difficult to enter
the OR command '||' in files. Since the inoremap setting was intended
for markdown files to auto-align inside tables on the | character, it is
better to let this remappping only apply in markdown files.

Also moved the complete-=t option to the new options file.

This change addresses the need by:

* Me